+++ tfjs-core/src/tensor.ts
@@ -301,7 +301,6 @@ export interface OpHandler {
       x: Tensor, begin: number[], end: number[], strides: number[],
       beginMask: number, endMask: number, ellipsisMask: number,
       newAxisMask: number, shrinkAxisMask: number): Tensor;
-  depthToSpace(x: Tensor4D, blockSize: number, dataFormat: string): Tensor4D;
   spectral: {
     fft(x: Tensor): Tensor; ifft(x: Tensor): Tensor; rfft(x: Tensor): Tensor;
     irfft(x: Tensor): Tensor
@@ -1194,12 +1193,6 @@ export class Tensor<R extends Rank = Rank> {
     return opHandler.stridedSlice(
         this, begin, end, strides, beginMask, endMask, ellipsisMask,
         newAxisMask, shrinkAxisMask);
-  }
-
-  depthToSpace(this: Tensor4D, blockSize: number, dataFormat: 'NHWC'|'NCHW'):
-      Tensor4D {
-    this.throwIfDisposed();
-    return opHandler.depthToSpace(this, blockSize, dataFormat);
   }
 
   fft(this: Tensor): Tensor {
